requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
bcrypt
scikit-learn
numpy